fix post-deploy script issue with shallow clones in CI
```
error: Could not find the SHA of the previous release in the git
history. If you limit the clone depth, try to increase it. Otherwise,
it means that the commit we are looking for was amended or squashed
and cannot be retrieved. Use --ignore-missing flag to skip it and
create a new release with the default commits count.
```

reference: https://docs.sentry.io/product/cli/releases/#dealing-with-missing-commits